refactor(pusher): address review nits
Remove the test-only lastError field and the isReadyForRetry helper,
neither of which had a production reader. Narrow the retry heap's
RWMutex to a plain Mutex, and rename retryHeap.Stop to Close since it
runs no background work. Route runTask panic recovery through the
injected logger instead of the stdlib log package.

Test pass: relocate retry-heap tests out of the batch and
state-callback files, replace sleep-based gates with require.Eventually,
add per-case subtests, and merge two overlapping expiry tests into one.

Raised by jefchien in review.
