+++ plugins/outputs/cloudwatchlogs/internal/pusher/retryheap_test.go
@@ -20,7 +20,7 @@ import (
 
 func TestRetryHeap(t *testing.T) {
 	heap := NewRetryHeap(&testutil.Logger{})
-	defer heap.Stop()
+	defer heap.Close()
 
 	// Test empty heap
 	assert.Equal(t, 0, heap.Size())
@@ -52,28 +52,26 @@ func TestRetryHeap(t *testing.T) {
 
 func TestRetryHeapOrdering(t *testing.T) {
 	heap := NewRetryHeap(&testutil.Logger{})
-	defer heap.Stop()
+	defer heap.Close()
 
 	target := Target{Group: "group", Stream: "stream"}
 	now := time.Now()
 
-	// Create batches with different retry times (not in order)
+	// Create batches with distinct retry times already in the past (not in order), so all
+	// are ready immediately and PopReady must return them ordered by nextRetryTime.
 	batch1 := newLogEventBatch(target, nil)
-	batch1.nextRetryTime = now.Add(30 * time.Millisecond)
+	batch1.nextRetryTime = now.Add(-10 * time.Millisecond)
 
 	batch2 := newLogEventBatch(target, nil)
-	batch2.nextRetryTime = now.Add(10 * time.Millisecond)
+	batch2.nextRetryTime = now.Add(-30 * time.Millisecond)
 
 	batch3 := newLogEventBatch(target, nil)
-	batch3.nextRetryTime = now.Add(20 * time.Millisecond)
+	batch3.nextRetryTime = now.Add(-20 * time.Millisecond)
 
 	// Push in random order
 	heap.Push(batch1)
 	heap.Push(batch2)
 	heap.Push(batch3)
-
-	// Wait for all to be ready
-	time.Sleep(100 * time.Millisecond)
 
 	// Pop ready batches - should come out in order
 	ready := heap.PopReady()
@@ -84,10 +82,10 @@ func TestRetryHeapOrdering(t *testing.T) {
 
 func TestRetryHeapProcessor(t *testing.T) {
 	heap := NewRetryHeap(&testutil.Logger{})
-	defer heap.Stop()
+	defer heap.Close()
 
 	// Create mock components with proper signature
-	workerPool := NewWorkerPool(2)
+	workerPool := NewWorkerPool(2, &testutil.Logger{})
 	defer workerPool.Stop()
 	mockService := &mockLogsService{}
 	mockTargetManager := &mockTargetManager{}
@@ -103,9 +101,9 @@ func TestRetryHeapProcessor(t *testing.T) {
 
 func TestRetryHeapProcessorExpiredBatch(t *testing.T) {
 	heap := NewRetryHeap(&testutil.Logger{})
-	defer heap.Stop()
+	defer heap.Close()
 
-	workerPool := NewWorkerPool(2)
+	workerPool := NewWorkerPool(2, &testutil.Logger{})
 	defer workerPool.Stop()
 	mockService := &mockLogsService{}
 	mockTargetManager := &mockTargetManager{}
@@ -114,8 +112,10 @@ func TestRetryHeapProcessorExpiredBatch(t *testing.T) {
 
 	target := Target{Group: "group", Stream: "stream"}
 	batch := newLogEventBatch(target, nil)
+	batch.append(newLogEvent(time.Now(), "test message", nil))
 	batch.initializeStartTime()
 	batch.expireAfter = time.Now().Add(-1 * time.Hour) // Already expired
+	batch.updateRetryMetadata(&cloudwatchlogs.ServiceUnavailableException{})
 	batch.nextRetryTime = time.Now().Add(-1 * time.Second)
 
 	var doneCalled, resumeCalled bool
@@ -123,18 +123,19 @@ func TestRetryHeapProcessorExpiredBatch(t *testing.T) {
 	batch.addResumeCallback(func() { resumeCalled = true })
 
 	heap.Push(batch)
+	assert.Equal(t, 1, heap.Size(), "batch should be queued before processing")
 
 	processor.processReadyMessages()
 	assert.Equal(t, 0, heap.Size(), "Expired batch should be removed from heap")
-	assert.True(t, resumeCalled, "expired batch should resume the circuit breaker")
+	assert.True(t, resumeCalled, "expired batch should resume the circuit breaker to unblock the target")
 	assert.False(t, doneCalled, "expired batch was never delivered, so it must not signal success")
 }
 
 func TestRetryHeapProcessorSendsBatch(t *testing.T) {
 	heap := NewRetryHeap(&testutil.Logger{})
-	defer heap.Stop()
+	defer heap.Close()
 
-	workerPool := NewWorkerPool(2)
+	workerPool := NewWorkerPool(2, &testutil.Logger{})
 	defer workerPool.Stop()
 
 	mockService := &mockLogsService{}
@@ -164,7 +165,7 @@ func TestRetryHeapProcessorSendsBatch(t *testing.T) {
 
 func TestRetryHeap_UnboundedPush(t *testing.T) {
 	heap := NewRetryHeap(&testutil.Logger{}) // maxSize parameter ignored (unbounded)
-	defer heap.Stop()
+	defer heap.Close()
 
 	// Push multiple batches without blocking
 	target := Target{Group: "group", Stream: "stream"}
@@ -207,9 +208,9 @@ func TestRetryHeap_UnboundedPush(t *testing.T) {
 
 func TestRetryHeapProcessorNoReadyBatches(t *testing.T) {
 	heap := NewRetryHeap(&testutil.Logger{})
-	defer heap.Stop()
+	defer heap.Close()
 
-	workerPool := NewWorkerPool(2)
+	workerPool := NewWorkerPool(2, &testutil.Logger{})
 	defer workerPool.Stop()
 	mockService := &mockLogsService{}
 	mockTargetManager := &mockTargetManager{}
@@ -224,9 +225,9 @@ func TestRetryHeapProcessorNoReadyBatches(t *testing.T) {
 
 func TestRetryHeapProcessorFailedBatchGoesBackToHeap(t *testing.T) {
 	heap := NewRetryHeap(&testutil.Logger{})
-	defer heap.Stop()
+	defer heap.Close()
 
-	workerPool := NewWorkerPool(2)
+	workerPool := NewWorkerPool(2, &testutil.Logger{})
 	defer workerPool.Stop()
 
 	// Create failing service with AWS error that triggers retry
@@ -266,8 +267,8 @@ func TestRetryHeapStopTwice(t *testing.T) {
 	rh := NewRetryHeap(&testutil.Logger{})
 
 	// Call Stop twice - should not panic
-	rh.Stop()
-	rh.Stop()
+	rh.Close()
+	rh.Close()
 
 	// After stopping, Push should drop the batch silently
 	target := Target{Group: "test-group", Stream: "test-stream"}
@@ -281,9 +282,9 @@ func TestRetryHeapStopTwice(t *testing.T) {
 
 func TestRetryHeapProcessorStoppedProcessReadyMessages(t *testing.T) {
 	heap := NewRetryHeap(&testutil.Logger{})
-	defer heap.Stop()
+	defer heap.Close()
 
-	workerPool := NewWorkerPool(2)
+	workerPool := NewWorkerPool(2, &testutil.Logger{})
 	defer workerPool.Stop()
 	mockService := &mockLogsService{}
 	mockTargetManager := &mockTargetManager{}
@@ -340,7 +341,7 @@ func TestRetryHeapProcessorStopDoesNotDeadlock(t *testing.T) {
 		return &cloudwatchlogs.PutLogEventsOutput{}, nil
 	})
 
-	workerPool := NewWorkerPool(1) // 1 worker + task buffer 2 -> Submit blocks on the 4th batch
+	workerPool := NewWorkerPool(1, &testutil.Logger{}) // 1 worker + task buffer 2 -> Submit blocks on the 4th batch
 	retryHeap := NewRetryHeap(logger)
 
 	// Saturate the pool first: the processor's own ticker drains these and parks on release.
@@ -381,7 +382,7 @@ func TestProcessorStopIsIdempotentAndConcurrencySafe(t *testing.T) {
 	service := okStubService(func(*cloudwatchlogs.PutLogEventsInput) (*cloudwatchlogs.PutLogEventsOutput, error) {
 		return &cloudwatchlogs.PutLogEventsOutput{}, nil
 	})
-	workerPool := NewWorkerPool(2)
+	workerPool := NewWorkerPool(2, &testutil.Logger{})
 	defer workerPool.Stop()
 	retryHeap := NewRetryHeap(logger)
 
@@ -411,12 +412,12 @@ func TestHeapPushAfterStopFailsCleanly(t *testing.T) {
 	logger := &testutil.Logger{}
 	h := NewRetryHeap(logger)
 	require.NoError(t, h.Push(readyBatch("g", nil, nil)))
-	h.Stop()
+	h.Close()
 
 	err := h.Push(readyBatch("g", nil, nil))
 	require.Error(t, err, "push after Stop must report failure, not silently succeed")
 	assert.Contains(t, err.Error(), "stopped")
-	h.Stop() // idempotent
+	h.Close() // idempotent
 }
 
 // RetryHeapProcessor.Stop() flushes ready batches through
@@ -429,7 +430,7 @@ func TestRetryHeapProcessorStopDoesNotBlockOnSaturatedPool(t *testing.T) {
 		return &cloudwatchlogs.PutLogEventsOutput{}, nil
 	})
 
-	pool := NewWorkerPool(1) // tasks buffer = size*2 = 2
+	pool := NewWorkerPool(1, &testutil.Logger{}) // tasks buffer = size*2 = 2
 	defer pool.Stop()
 
 	// Occupy the single worker and fill the task buffer so any further Submit blocks.
@@ -441,7 +442,7 @@ func TestRetryHeapProcessorStopDoesNotBlockOnSaturatedPool(t *testing.T) {
 	pool.Submit(func() { <-release })
 
 	retryHeap := NewRetryHeap(logger)
-	defer retryHeap.Stop()
+	defer retryHeap.Close()
 	p := NewRetryHeapProcessor(retryHeap, pool, svc, NewTargetManager(logger, svc), logger, nil)
 
 	// Ready batches force Stop()'s flush to Submit into the saturated pool.
@@ -479,7 +480,7 @@ func TestProcessorStopWhenProcessLoopBlockedInSubmit(t *testing.T) {
 		return &cloudwatchlogs.PutLogEventsOutput{}, nil
 	})
 
-	pool := NewWorkerPool(1) // task buffer = 2
+	pool := NewWorkerPool(1, &testutil.Logger{}) // task buffer = 2
 	defer pool.Stop()
 	release := make(chan struct{})
 	pool.Submit(func() { <-release }) // occupy the worker
@@ -487,7 +488,7 @@ func TestProcessorStopWhenProcessLoopBlockedInSubmit(t *testing.T) {
 	pool.Submit(func() { <-release }) // fill buffer
 
 	heap := NewRetryHeap(logger)
-	defer heap.Stop()
+	defer heap.Close()
 	p := NewRetryHeapProcessor(heap, pool, svc, NewTargetManager(logger, svc), logger, nil)
 	require.NoError(t, heap.Push(readyBatch("g", nil, nil)))
 
@@ -505,4 +506,203 @@ func TestProcessorStopWhenProcessLoopBlockedInSubmit(t *testing.T) {
 			"and the pool is only stopped after Stop() returns, so nothing releases it")
 	}
 	close(release)
+}
+
+// A batch dropped because the retry heap is already stopped (shutdown in
+// progress) must not fire done callbacks. done() runs updateState() AND every
+// doneCallback -- per-event LogEvent.Done() plus the queue's onSuccessCallback -- which
+// marks never-delivered events as delivered and advances file offsets past them, so the
+// events are neither shipped nor re-read after restart.
+func TestHeapStoppedDropDoesNotSignalSuccess(t *testing.T) {
+	logger := &testutil.Logger{}
+	service := okStubService(func(*cloudwatchlogs.PutLogEventsInput) (*cloudwatchlogs.PutLogEventsOutput, error) {
+		return nil, &cloudwatchlogs.ServiceUnavailableException{}
+	})
+
+	retryHeap := NewRetryHeap(logger)
+	retryHeap.Close() // shutdown already closed the heap
+
+	var delivered, offsetsAdvanced atomic.Int32
+	batch := readyBatch("g", func() { delivered.Add(1) }, func() { offsetsAdvanced.Add(1) })
+
+	newSender(logger, service, NewTargetManager(logger, service), retryHeap).Send(batch)
+
+	assert.Zero(t, delivered.Load(),
+		"undelivered batch dropped at shutdown fired done callbacks: events are reported delivered")
+	assert.Zero(t, offsetsAdvanced.Load(),
+		"undelivered batch dropped at shutdown advanced file offsets: events are lost across restart")
+}
+
+// Second site: the same false-success signal on the expired-batch drop
+// inside flushReadyBatches. Expiry is a permanent give-up, so offsets SHOULD advance --
+// but done callbacks must not fire, because nothing was delivered.
+func TestExpiredBatchDropDoesNotSignalSuccess(t *testing.T) {
+	logger := &testutil.Logger{}
+	service := okStubService(func(*cloudwatchlogs.PutLogEventsInput) (*cloudwatchlogs.PutLogEventsOutput, error) {
+		return &cloudwatchlogs.PutLogEventsOutput{}, nil
+	})
+
+	workerPool := NewWorkerPool(1, &testutil.Logger{})
+	defer workerPool.Stop()
+	retryHeap := NewRetryHeap(logger)
+
+	var delivered, offsetsAdvanced atomic.Int32
+	batch := readyBatch("g", func() { delivered.Add(1) }, func() { offsetsAdvanced.Add(1) })
+	batch.expireAfter = time.Now().Add(-time.Hour) // already past its expiry window
+	require.NoError(t, retryHeap.Push(batch))
+
+	p := NewRetryHeapProcessor(retryHeap, workerPool, service, NewTargetManager(logger, service), logger, nil)
+	p.flushReadyBatches()
+
+	require.True(t, batch.isExpired(), "test setup: batch must be expired to exercise the drop path")
+	assert.Zero(t, delivered.Load(),
+		"expired batch fired done callbacks: never-delivered events are reported delivered")
+	assert.Equal(t, int32(1), offsetsAdvanced.Load(),
+		"expired batch is a permanent give-up, so offsets should advance exactly once")
+}
+
+// TestRetryHeapSuccessCallsStateCallback verifies that when a batch succeeds
+// on retry through the heap, state callbacks fire to persist file offsets.
+func TestRetryHeapSuccessCallsStateCallback(t *testing.T) {
+	logger := &testutil.Logger{}
+	target := Target{Group: "group", Stream: "stream"}
+
+	queue := &mockFileRangeQueue{}
+	queue.On("ID").Return("file1")
+	queue.On("Enqueue", mock.Anything).Return()
+
+	service := &stubLogsService{
+		ple: func(_ *cloudwatchlogs.PutLogEventsInput) (*cloudwatchlogs.PutLogEventsOutput, error) {
+			return &cloudwatchlogs.PutLogEventsOutput{}, nil
+		},
+		cls: func(_ *cloudwatchlogs.CreateLogStreamInput) (*cloudwatchlogs.CreateLogStreamOutput, error) {
+			return &cloudwatchlogs.CreateLogStreamOutput{}, nil
+		},
+		clg: func(_ *cloudwatchlogs.CreateLogGroupInput) (*cloudwatchlogs.CreateLogGroupOutput, error) {
+			return &cloudwatchlogs.CreateLogGroupOutput{}, nil
+		},
+		dlg: func(_ *cloudwatchlogs.DescribeLogGroupsInput) (*cloudwatchlogs.DescribeLogGroupsOutput, error) {
+			return &cloudwatchlogs.DescribeLogGroupsOutput{}, nil
+		},
+	}
+
+	retryHeap := NewRetryHeap(logger)
+	workerPool := NewWorkerPool(2, &testutil.Logger{})
+	tm := NewTargetManager(logger, service)
+	defer retryHeap.Close()
+	defer workerPool.Stop()
+
+	processor := NewRetryHeapProcessor(retryHeap, workerPool, service, tm, logger, retryer.NewLogThrottleRetryer(logger))
+
+	batch := newStatefulBatch(target, queue)
+	batch.nextRetryTime = time.Now().Add(-1 * time.Second)
+
+	err := retryHeap.Push(batch)
+	assert.NoError(t, err)
+
+	processor.processReadyMessages()
+	time.Sleep(200 * time.Millisecond)
+
+	assert.Equal(t, 0, retryHeap.Size(), "Heap should be empty after success")
+	queue.AssertCalled(t, "Enqueue", mock.Anything)
+}
+
+// TestRetryHeapExpiryCallsStateCallback verifies that when a batch expires
+// after 14 days without successfully publishing, state callbacks still fire
+// to persist file offsets and prevent re-reading on restart.
+func TestRetryHeapExpiryCallsStateCallback(t *testing.T) {
+	logger := &testutil.Logger{}
+	target := Target{Group: "group", Stream: "stream"}
+
+	queue := &mockFileRangeQueue{}
+	queue.On("ID").Return("file1")
+	queue.On("Enqueue", mock.Anything).Return()
+
+	service := &stubLogsService{
+		ple: func(_ *cloudwatchlogs.PutLogEventsInput) (*cloudwatchlogs.PutLogEventsOutput, error) {
+			return nil, &cloudwatchlogs.ServiceUnavailableException{}
+		},
+		cls: func(_ *cloudwatchlogs.CreateLogStreamInput) (*cloudwatchlogs.CreateLogStreamOutput, error) {
+			return &cloudwatchlogs.CreateLogStreamOutput{}, nil
+		},
+		clg: func(_ *cloudwatchlogs.CreateLogGroupInput) (*cloudwatchlogs.CreateLogGroupOutput, error) {
+			return &cloudwatchlogs.CreateLogGroupOutput{}, nil
+		},
+		dlg: func(_ *cloudwatchlogs.DescribeLogGroupsInput) (*cloudwatchlogs.DescribeLogGroupsOutput, error) {
+			return &cloudwatchlogs.DescribeLogGroupsOutput{}, nil
+		},
+	}
+
+	retryHeap := NewRetryHeap(logger)
+	workerPool := NewWorkerPool(2, &testutil.Logger{})
+	tm := NewTargetManager(logger, service)
+	defer retryHeap.Close()
+	defer workerPool.Stop()
+
+	processor := NewRetryHeapProcessor(retryHeap, workerPool, service, tm, logger, nil)
+
+	batch := newStatefulBatch(target, queue)
+	batch.initializeStartTime()
+	batch.expireAfter = time.Now().Add(-10 * time.Millisecond) // Already expired
+	batch.updateRetryMetadata(&cloudwatchlogs.ServiceUnavailableException{})
+	batch.nextRetryTime = time.Now().Add(-1 * time.Second) // Override to make it ready
+
+	err := retryHeap.Push(batch)
+	assert.NoError(t, err)
+
+	processor.processReadyMessages()
+	time.Sleep(200 * time.Millisecond)
+
+	assert.Equal(t, 0, retryHeap.Size(), "Expired batch should be removed")
+	queue.AssertCalled(t, "Enqueue", mock.Anything)
+}
+
+// TestShutdownDoesNotCallStateCallback verifies that during a clean shutdown
+// via Stop(), remaining batches in the retry heap do NOT have their state
+// callbacks invoked. This prevents marking undelivered data as processed.
+func TestShutdownDoesNotCallStateCallback(t *testing.T) {
+	logger := &testutil.Logger{}
+	target := Target{Group: "group", Stream: "stream"}
+
+	var stateCallCount atomic.Int32
+
+	retryHeap := NewRetryHeap(logger)
+	workerPool := NewWorkerPool(2, &testutil.Logger{})
+	defer workerPool.Stop()
+
+	service := &stubLogsService{
+		ple: func(_ *cloudwatchlogs.PutLogEventsInput) (*cloudwatchlogs.PutLogEventsOutput, error) {
+			return nil, &cloudwatchlogs.ServiceUnavailableException{}
+		},
+		cls: func(_ *cloudwatchlogs.CreateLogStreamInput) (*cloudwatchlogs.CreateLogStreamOutput, error) {
+			return &cloudwatchlogs.CreateLogStreamOutput{}, nil
+		},
+		clg: func(_ *cloudwatchlogs.CreateLogGroupInput) (*cloudwatchlogs.CreateLogGroupOutput, error) {
+			return &cloudwatchlogs.CreateLogGroupOutput{}, nil
+		},
+		dlg: func(_ *cloudwatchlogs.DescribeLogGroupsInput) (*cloudwatchlogs.DescribeLogGroupsOutput, error) {
+			return &cloudwatchlogs.DescribeLogGroupsOutput{}, nil
+		},
+	}
+	tm := NewTargetManager(logger, service)
+
+	processor := NewRetryHeapProcessor(retryHeap, workerPool, service, tm, logger, nil)
+	processor.Start()
+
+	// Push a batch with a future retry time so it won't be processed before Stop
+	batch := newLogEventBatch(target, nil)
+	batch.append(newLogEvent(time.Now(), "test", nil))
+	batch.addStateCallback(func() { stateCallCount.Add(1) })
+	batch.nextRetryTime = time.Now().Add(1 * time.Hour) // Not ready yet
+
+	err := retryHeap.Push(batch)
+	assert.NoError(t, err)
+
+	// Stop the processor — batch is still in heap, not ready
+	processor.Stop()
+	retryHeap.Close()
+
+	assert.Equal(t, int32(0), stateCallCount.Load(),
+		"State callback should not be called for unprocessed batches during shutdown")
+	assert.Equal(t, 1, retryHeap.Size(), "Batch should remain in heap after shutdown")
 }
